bugfix - OptionItem - set Weight to Decimal (#324)

+++ code/objects/OptionItem.php
@@ -10,7 +10,7 @@ class OptionItem extends DataObject
      */
     private static $db = array(
         'Title' => 'Text',
-        'WeightModifier' => 'Int',
+        'WeightModifier' => 'Decimal',
         'CodeModifier' => 'Text',
         'PriceModifier' => 'Currency',
         'WeightModifierAction' => "Enum('Add,Subtract,Set','Add')",
